move results out of init
